backend/requirements.txt
fastapi
uvicorn[standard]
websockets
httpx
pyzeebe
python-dotenv
aiofiles
PyJWT
